requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.4.1

# Vector Database
qdrant-client==1.7.3

# Authentication
python-jose[cryptography]==3.3.0

# Environment
python-dotenv==1.0.0

# OpenAI API
openai==2.7.1

# Azure Speech Services
azure-cognitiveservices-speech==1.34.0

# Document Parsing
PyPDF2==3.0.1
python-docx==1.1.0
chardet==5.2.0
beautifulsoup4==4.12.3
PyJWT==2.10.1

# Audio Processing
pydub==0.25.1
requests==2.31.0

# HTTP Client
aiohttp==3.9.1
